+++ pdf_viewer/main.cpp
@@ -714,10 +714,10 @@ int main(int argc, char* args[]) {
 	quit = true;
 
 	std::vector<MainWidget*> windows_to_delete;
-	for (int i = 0; i < windows.size(); i++) {
+	for (size_t i = 0; i < windows.size(); i++) {
 		windows_to_delete.push_back(windows[i]);
 	}
-	for (int i = 0; i < windows_to_delete.size(); i++) {
+	for (size_t i = 0; i < windows_to_delete.size(); i++) {
 		delete windows_to_delete[i];
 	}
